Mastery Recursion Fix, Level Cap Compatibility
Removed all "isLevelMax" checks and replaced them with their default max config value.

Fix for recursion in AncientSwordMastery by moving the event from LivingDamageEvent to LivingHurtEvent.
This'll definitely change the behavior of it as an enchantment, but at least now it probably won't crash.
It may be possible to do a better fix later.